add view models and groups feature
